move persistence versioning tests into own file
