+++ test/server/data/persistence/persistence.spec.js
@@ -356,99 +356,6 @@ describe('Server', () => {
             });
         });
       });
-
-      describe('versioning', () => {
-        const userId = testUserId;
-
-        let versionLog;
-        let versions;
-        const nonExistentSHA = '795c5751c8e0b0c9b5993ec81928cd89f7eefd27';
-        const projectData = new Project(updateProjectWithAuthor());
-        const projectId = projectData.id;
-        const projectRepoDataPath = filePaths.createProjectDataPath(projectId);
-        const newProject = projectData.merge({ projectData: 'new stuff' });
-
-        const blockData = Block.classless({ projectId });
-        const blockId = blockData.id;
-        const newBlock = merge({}, blockData, { blockData: 'new data' });
-
-        //skip test suite if not using s3
-        before(function () {
-          if (s3.useRemote) {
-            this.skip();
-          }
-
-          return persistence.projectCreate(projectId, projectData, userId) //3
-            .then(() => persistence.blockWrite(projectId, blockData))
-            .then(() => persistence.projectSave(projectId, userId)) //2
-            .then(() => persistence.projectWrite(projectId, newProject, userId))
-            .then(() => persistence.projectSave(projectId, userId)) //1
-            .then(() => persistence.blockWrite(projectId, newBlock))
-            .then(() => persistence.projectSave(projectId, userId)) //0
-            .then(() => versioning.log(projectRepoDataPath))
-            .then(log => {
-              versionLog = log;
-              versions = versionLog.map(commit => commit.sha);
-            });
-        });
-
-        it('projectExists() rejects if invalid version', () => {
-          return persistence.projectExists(projectId, 'invalidSHA')
-            .then(result => assert(false, 'should not resolve'))
-            .catch(err => assert(err === errorDoesNotExist, 'wrong error type -> function errored...'));
-        });
-
-        it('projectExists() rejects if version does not exist', () => {
-          return persistence.projectExists(projectId, nonExistentSHA)
-            .then(result => assert(false, 'should not resolve'))
-            .catch(err => assert(err === errorDoesNotExist, 'wrong error type -> function errored...'));
-        });
-
-        it('projectExists() resolves if version exists', () => {
-          return persistence.projectExists(projectId, versions[1]);
-        });
-
-        it('projectGet() rejects on if given invalid version', () => {
-          return persistence.projectGet(projectId, nonExistentSHA)
-            .then(result => assert(false, 'should not resolve'))
-            .catch(err => assert(err === errorDoesNotExist, 'wrong error type -> function errored...'));
-        });
-
-        it('projectGet() resolves to correct file when given version', () => {
-          return persistence.projectGet(projectId, versions[2])
-            .then(project => expect(project).to.eql(projectData));
-        });
-
-        it('projectGet() defaults to latest version', () => {
-          return persistence.projectGet(projectId)
-            .then(project => expect(project).to.eql(Object.assign({}, newProject, {
-              version: versions[0],
-              lastSaved: versionLog[0].time
-            })));
-        });
-
-        it('blockExists() rejects on invalid version', () => {
-          return persistence.blocksExist(projectId, nonExistentSHA, blockId)
-            .then(result => assert(false, 'should not resolve'))
-            .catch(err => assert(err === errorDoesNotExist, 'wrong error type -> function errored...'));
-        });
-
-        it('blockExists() accepts a version', () => {
-          return persistence.blocksExist(projectId, versions[2], blockId);
-        });
-
-        it('blockGet() accepts a version, gets version at that time', () => {
-          return persistence.blocksGet(projectId, versions[2], blockId)
-            .then(blockMap => blockMap[blockId])
-            .then(block => expect(block).to.eql(blockData));
-        });
-
-        it('blockGet() defaults to latest version', () => {
-          return persistence.blocksGet(projectId, false, blockId)
-            .then(blockMap => blockMap[blockId])
-            .then(block => expect(block).to.eql(newBlock));
-        });
-      });
     });
   });
 });
